Move work gen mocks to mirrmock module
