+++ mirrulations-work-generator/tests/test_workgen.py
@@ -2,8 +2,8 @@ from fakeredis import FakeRedis
 from mirrgen.job_queue import JobQueue
 from mirrgen.work_generator import WorkGenerator
 from mirrgen.regulations_api import RegulationsAPI
-from mirrgen.mock_dataset import MockDataSet
-from mirrgen.mock_data_storage import MockDataStorage
+from mirrmock.mock_dataset import MockDataSet
+from mirrmock.mock_data_storage import MockDataStorage
 
 
 def test_work_generator_single_page(requests_mock, mocker):
